Use getentropy() directly
In 2024, getentropy() is effectively universal but the arc4random family
of functions is not. A correct userspace arc4random generator can be
lengthy as it should be thread-safe and needs to detect forks to avoid
reuse of cloned state.

Luckily, we make infrequent, unsophisticated use of randomness, so it is
fine to inline a handful of direct getentropy() calls.

Include some preemptive comments to fend off complaints that, unlike
arc4random_uniform(), simple modular reduction from 2**64 without rejection
is very slightly non-uniform in [0, 3000). It doesn't matter.