Add -s and -S to require or prohibit setting time
If -s is given, ntpd will exit with an error if it fails to set the time on
startup in the foreground. Constraints, trusted peers or trusted sensors
are still required to do this.

If -S is given, no attempt is made to set the time and ntpd immediately
forks into the background to continue normal operation.

Drop the automatic behaviour change depending on BSD securelevel. Invoke
with an explicit -S to disable attempts to set the clock in secure mode.