Replace ntpctl with ntpq
For reasons lost in the mists of time, ntpctl has a horrible interface with
a mandatory -s option taking a case-sensitive keyword indicating what kind
of status info to request. No other arguments or options are supported.

Replace ntpctl with ntpq which takes a single optional keyword. Run without
any arguments, it defaults to 'all' which is invariably what you want.

The program name is already included in the output of errx() so don't
repeat it in the message body.