Add initproctitle() and setsync() hooks
Platforms without a native setproctitle() need to clone argument and
environment strings before they are used to make space for the replacement
process title at the original argv[0] address. Add an initproctitle() hook
in the right place for this. Because the system __progname may point into
argv[0], we need to set __progname ourselves to avoid it being overwritten.

Some platforms need ntpd to update the sync status of the clock. Add a
setsync() hook in the right place for this.