Add a Makefile compatible with both BSD and GNU make
There's enough common ground between modern GNU and BSD make to allow us to
write a Makefile that suits both, although the != shell assignment operator
we use is not available in antique versions of GNU make.

Define the paths and ntp username in the Makefile so they can be overridden
when invoking make rather than hard-coding them in ntpd.h.