Add FreeBSD and DragonFly compatibility layer
DragonFly is closely related to FreeBSD with excellent cross-compatibility
so the two platforms can share a shim directory.

Implement adjfreq() and setsync() in terms of native ntp_adjtime() and
use the libc adjtime() directly.

BSDs have a native setproctitle() but we can use the initproctitle()
hook to canonicalise argv[0] using sysctl() with KERN_PROC_PATHNAME,
relaxing the requirement to run ntpd with an absolute pathname.