Ensure kernel sync status is always updated
The kernel's adjtimex() sync status needs to be updated whenever the
ntp client's sync status is changed, not just as a side-effect of
ntpd_adjtime(). In particular, STA_UNSYNC should be set at startup and
when no peers are replying.

Communicate this from the ntp engine back to the privileged main process
with IMSG_UNSYNCED, which is already used to signal the situation to the
dns engine. As an optimisation, also continue to set the sync status as
a side-effect of IMSG_ADJTIME: we would request this unconditionally on
receipt of an IMSG_ADJTIME reply anyway.